preprocessor/requirements.txt
pika==1.3.2
PyYAML==6.0.1
numpy==1.25.2
SQLAlchemy==1.4.49
mysql-connector-python==8.1.0
rouge==1.0.1
sentence-transformers==2.2.2
torch==2.1.0
transformers==4.34.0
konlpy==0.6.0